Move AppReadyHookArgs to web-pkg
