+++ packages/web-pkg/src/apps/index.ts
@@ -1,0 +1,1 @@
+export * from './types'
